+++ controllers/messages.controllers.js
@@ -1,6 +1,7 @@
 const MessagesModel = require('../model/Messages.model');
+const mongoose = require('mongoose')
 
-exports.getAll = async (req, res) => {
+exports.getAll = async (req, res, next) => {
 	try {
 		const { page = 1, limit } = req.query;
 		const response = await MessagesModel.find()
@@ -11,11 +12,11 @@ exports.getAll = async (req, res) => {
 		const pages = limit === undefined ? 1 : Math.ceil(total / limit);
 		res.json({ total, pages, status: 200, response });
 	} catch (error) {
-		res.json({ status: 404, message: error });
+		next({ status: 404, message: error });
 	}
 };
 
-exports.getWithQuery = async (req, res) => {
+exports.getWithQuery = async (req, res, next) => {
 	try {
 		const query =
 			typeof req.body.query === 'string'
@@ -35,11 +36,11 @@ exports.getWithQuery = async (req, res) => {
 			response,
 		});
 	} catch (error) {
-		res.json({ status: 404, message: error });
+		next({ status: 404, message: error });
 	}
 };
 
-exports.create = async (req, res) => {
+exports.create = async (req, res, next) => {
 	const {
 		firstname,
 		lastname,
@@ -74,27 +75,41 @@ exports.create = async (req, res) => {
 				response,
 			})
 		)
-		.catch((err) => res.json({ status: 404, message: err }));
+		.catch((err) => next({ status: 400, message: err }));
 };
 
-exports.getSingleMessage = async (req, res) => {
-	await MessagesModel.findById({ _id: req.params.id }, (err, data) => {
-		if (err) {
-			res.json({ status: 404, message: err });
-		} else {
-			res.json({ status: 200, data });
-		}
-	});
+exports.getSingleMessage = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.id)) {
+		await MessagesModel.findById({_id: req.params.id})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Messages Model.',
+					})
+				} else {
+					await MessagesModel.findById({ _id: req.params.id }, (err, data) => {
+						if (err) {
+							next({ status: 404, message: err });
+						} else {
+							res.json({ status: 200, data });
+						}
+					});
+				}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
 
-exports.getMessagesBySubject = async (req, res) => {
+exports.getMessagesBySubject = async (req, res, next) => {
 	const { page, limit } = req.query;
 	const total = await MessagesModel.find().countDocuments();
 	const pages = limit === undefined ? 1 : Math.ceil(total / limit);
 
 	await MessagesModel.find({ subject: req.params.subject }, (err, data) => {
 		if (err) {
-			res.json({ status: 404, message: err });
+			next({ status: 404, message: err });
 		} else {
 			res.json({ total, pages, status: 200, data });
 		}
@@ -104,14 +119,14 @@ exports.getMessagesBySubject = async (req, res) => {
 		.sort({ createdAt: -1 });
 };
 
-exports.getMessagesByEmail = async (req, res) => {
+exports.getMessagesByEmail = async (req, res, next) => {
 	const { page, limit } = req.query;
 	const total = await MessagesModel.find().countDocuments();
 	const pages = limit === undefined ? 1 : Math.ceil(total / limit);
 
 	await MessagesModel.find({ email: req.params.email }, (err, data) => {
 		if (err) {
-			res.json({ status: 404, message: err });
+			next({ status: 404, message: err });
 		} else {
 			res.json({ total, pages, status: 200, data });
 		}
@@ -121,18 +136,46 @@ exports.getMessagesByEmail = async (req, res) => {
 		.sort({ createdAt: -1 });
 };
 
-exports.updateMessage = async (req, res) => {
-	await MessagesModel.findByIdAndUpdate({ _id: req.params.id }, { $set: req.body })
-		.then((data) =>
-			res.json({ status: 200, message: 'Message is updated successfully', data })
-		)
-		.catch((err) => res.json({ status: 404, message: err }));
+exports.updateMessage = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.id)) {
+		await MessagesModel.findById({_id: req.params.id})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Messages Model.',
+					})
+				} else {
+					await MessagesModel.findByIdAndUpdate({ _id: req.params.id }, { $set: req.body })
+					.then((data) =>
+						res.json({ status: 200, message: 'Message is updated successfully', data })
+					)
+					.catch((err) => next({ status: 404, message: err }));
+				}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
 
-exports.removeSingleMessage = async (req, res) => {
-	await MessagesModel.findByIdAndDelete({ _id: req.params.id })
-		.then((data) =>
-			res.json({ status: 200, message: 'Message is deleted successfully', data })
-		)
-		.catch((err) => res.json({ status: 404, message: err }));
+exports.removeSingleMessage = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.id)) {
+		await MessagesModel.findById({_id: req.params.id})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Messages Model.',
+					})
+				} else {
+					await MessagesModel.findByIdAndDelete({ _id: req.params.id })
+					.then((data) =>
+						res.json({ status: 200, message: 'Message is deleted successfully', data })
+					)
+					.catch((err) => next({ status: 404, message: err }));
+				}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
