+++ controllers/roles.controllers.js
@@ -1,6 +1,7 @@
 const RolesModel = require('../model/Roles.model');
+const mongoose = require('mongoose')
 
-exports.getAllRoles = async (req, res) => {
+exports.getAllRoles = async (req, res, next) => {
 	try {
 		const { page = 1, limit } = req.query;
 		const total = await RolesModel.find().countDocuments();
@@ -11,23 +12,38 @@ exports.getAllRoles = async (req, res) => {
 			.sort({ createdAt: -1 });
 		res.json({ total, pages, status: 200, response });
 	} catch (error) {
-		res.json({ status: 404, message: error });
+		next({ status: 404, message: error });
 	}
 };
 
-exports.getSingleRole = async (req, res) => {
-	await RolesModel.findById({ _id: req.params.roleid }, (err, data) => {
-		if (err) {
-			res.json({ status: 404, message: err });
-		} else {
-			res.json({ status: 200, data });
-		}
-	});
+exports.getSingleRole = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.roleid)) {
+		await RolesModel.findById({_id: req.params.roleid})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Roles Model.',
+					})
+				} else {
+					await RolesModel.findById({ _id: req.params.roleid }, (err, data) => {
+						if (err) {
+							next({ status: 404, message: err });
+						} else {
+							res.json({ status: 200, data });
+						}
+					});
+				}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
 
-exports.createRole = async (req, res) => {
+exports.createRole = async (req, res, next) => {
 	const newRole = await new RolesModel({
 		name: req.body.name,
+		delete: req.body.delete
 	});
 	newRole
 		.save()
@@ -38,21 +54,49 @@ exports.createRole = async (req, res) => {
 				response,
 			})
 		)
-		.catch((err) => res.json({ status: 404, message: err }));
+		.catch((err) => next({ status: 400, message: err }));
 };
 
-exports.updateRole = async (req, res) => {
-	await RolesModel.findByIdAndUpdate({ _id: req.params.roleid }, { $set: req.body })
-		.then((data) =>
-			res.json({ status: 200, message: 'Role is updated successfully', data })
-		)
-		.catch((err) => res.json({ status: 404, message: err }));
+exports.updateRole = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.roleid)) {
+		await RolesModel.findById({_id: req.params.roleid})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Roles Model.',
+					})
+				} else {
+					await RolesModel.findByIdAndUpdate({ _id: req.params.roleid }, { $set: req.body })
+					.then((data) =>
+						res.json({ status: 200, message: 'Role is updated successfully', data })
+					)
+					.catch((err) => next({ status: 404, message: err }));
+				}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
 
-exports.removeRole = async (req, res) => {
-	await RolesModel.findByIdAndDelete({ _id: req.params.roleid })
-		.then((data) =>
-			res.json({ status: 200, message: 'Role is deleted successfully', data })
-		)
-		.catch((err) => res.json({ status: 404, message: err }));
+exports.removeRole = async (req, res, next) => {
+	if(mongoose.isValidObjectId(req.params.roleid)) {
+		await RolesModel.findById({_id: req.params.roleid})
+			.then(async(isExist) => {
+				if(isExist === null) {
+					next({
+						status: 404,
+						message: 'This Id is not exist in Roles Model.',
+					})
+				} else {
+					await RolesModel.findByIdAndDelete({ _id: req.params.roleid })
+					.then((data) =>
+						res.json({ status: 200, message: 'Role is deleted successfully', data })
+					)
+					.catch((err) => next({ status: 404, message: err }));
+							}
+			}).catch(err => next({status: 500, message:err}))
+	} else {
+		next({ status: 400, message: 'Object Id is not valid.' })
+	}
 };
